delete unused shacl validators
