+++ amf-client/shared/src/test/scala/amf/client/validation/ClientPayloadValidationTest.scala
